drop the composite pattern
