+++ cibo/actions/__action__.py
@@ -10,6 +10,7 @@ from cibo.password import Password
 from cibo.resources.doors import Doors
 from cibo.resources.items import Items
 from cibo.resources.npcs import Npcs
+from cibo.resources.resources import Resources
 from cibo.resources.rooms import Rooms
 
 
@@ -28,6 +29,17 @@ class Action(ABC):
         self._output = self._server_config.output
 
         self._password_hasher = Password()
+
+    @property
+    def resources(self) -> Resources:
+        """Resource helper methods that aren't necesarily associated with just one
+        resource type.
+
+        Returns:
+            Resources: The helper methods.
+        """
+
+        return self._world.resources
 
     @property
     def rooms(self) -> Rooms:
