+++ cibo/resources/items.py
@@ -10,7 +10,7 @@ from cibo.exception import ItemNotFound
 from cibo.models.data.item import Item as ItemData
 from cibo.models.description import EntityDescription
 from cibo.models.item import Item
-from cibo.resources.__resource__ import Composite, Resource
+from cibo.resources.__resource__ import Resource
 
 
 class Items(Resource):
@@ -18,7 +18,6 @@ class Items(Resource):
 
     def __init__(self, items_file: str):
         self._items: List[Item] = self._generate_resources(items_file)
-        self.get_by_name = Composite().get_by_name
 
     def _create_resource_from_dict(self, resource: dict) -> Item:
         item = resource
