+++ cibo/resources/npcs.py
@@ -10,7 +10,7 @@ from cibo.exception import NpcNotFound
 from cibo.models.data.npc import Npc as NpcData
 from cibo.models.description import EntityDescription
 from cibo.models.npc import Npc
-from cibo.resources.__resource__ import Composite, Resource
+from cibo.resources.__resource__ import Resource
 
 
 class Npcs(Resource):
@@ -18,7 +18,6 @@ class Npcs(Resource):
 
     def __init__(self, npcs_file: str):
         self._npcs: List[Npc] = self._generate_resources(npcs_file)
-        self.get_by_name = Composite().get_by_name
 
     def _create_resource_from_dict(self, resource: dict) -> Npc:
         npc = resource
