+++ cibo/resources/world.py
@@ -9,6 +9,7 @@ from cibo.resources.doors import Doors
 from cibo.resources.items import Items
 from cibo.resources.npcs import Npcs
 from cibo.resources.regions import Regions
+from cibo.resources.resources import Resources
 from cibo.resources.rooms import Rooms
 from cibo.resources.sectors import Sectors
 from cibo.resources.spawns import Spawns
@@ -20,6 +21,8 @@ class World:
     """
 
     def __init__(self) -> None:
+        self.resources = Resources()
+
         self.regions = Regions(getenv("REGIONS_PATH", "/cibo/config/regions.json"))
         self.sectors = Sectors(
             getenv("SECTORS_PATH", "/cibo/config/sectors.json"), self.regions
